aider: Added a new block for calling a function in `custom_blocks.js` and added the new block to the toolbox in `index.html`.
# Aider chat conversation:

ASSISTANT: To add a new block for calling a function, we need to modify the `custom_blocks.js` and `index.html` files.

In the `custom_blocks.js` file, we need to define a new block type, let's call it `functionCall`. This block will have a single field for the function name.

Here is the change:

```javascript
Definitions/custom_blocks.js
<<<<<<< HEAD
=======
Blockly.Blocks['functionCall'] = {
  init: function() {
    this.appendDummyInput()
        .appendField(new Blockly.FieldTextInput("f"), "NAME");
    this.setColour(20);
    this.setTooltip("");
    this.setHelpUrl("");
  }
};

Blockly.JavaScript['functionCall'] = function(block) {
  var func_name = block.getFieldValue('NAME');
  var code = func_name + '()';
  return [code, Blockly.JavaScript.ORDER_FUNCTION_CALL];
};
>>>>>>> updated
```

In the `index.html` file, we need to add the new block to the toolbox.

Here is the change:

```html
Definitions/index.html
<<<<<<< HEAD
        <xml id="toolbox" style="display: none">
            <block type="functionDef"></block>
            <block type="abs"></block>
            <block type="app"></block>
            <block type="var"></block>
            <block type="math_number"></block>
            <block type="math_arithmetic"></block>
        </xml>
=======
        <xml id="toolbox" style="display: none">
            <block type="functionDef"></block>
            <block type="functionCall"></block>
            <block type="abs"></block>
            <block type="app"></block>
            <block type="var"></block>
            <block type="math_number"></block>
            <block type="math_arithmetic"></block>
        </xml>
>>>>>>> updated
```

These changes will add a new block for calling a function. The block will have the same color as the function definition block and will be placed right after it in the toolbox.