aider: Increased the height of the window with the toolbox and the workspace to take up 75% of the viewport.
# Aider chat conversation:

USER: I want to make the window with the toolbox and the workspace bigger, increasing the height.

ASSISTANT: To increase the height of the window with the toolbox and the workspace, we need to modify the CSS rules in the `styles.css` file. Specifically, we need to change the `height` property of the `.row` class.

Here is the *edit block* for the change:

```css
Definitions/styles.css
<<<<<<< HEAD
.row {
    display: flex;
    justify-content: space-between;
    height: 50vh; /* Take up half the height of the viewport */
}
=======
.row {
    display: flex;
    justify-content: space-between;
    height: 75vh; /* Take up 75% of the height of the viewport */
}
>>>>>>> updated
```

This change will make the `.row` class take up 75% of the height of the viewport, instead of 50%. Adjust the value as needed to get the desired height.