aider: Added a link to the GitHub source code in the HTML file.
# Aider chat conversation:

USER: I want to add a link to the github source code that can be clicked upon when looking at this on Github-Pages

ASSISTANT: To add a link to the GitHub source code, we can add an anchor tag (`<a>`) in the HTML file. This tag will be placed in the `body` of the HTML file, preferably at the bottom or top of the page, depending on your preference.

Here is the change that needs to be made:

```html
Definitions/index.html
<<<<<<< HEAD
    <div class="row">
        <div id="blocklyDiv" class="full-width"></div>
        <xml id="toolbox" style="display: none">
            <block type="functionDef"></block>
            <block type="functionCall"></block>
            <block type="abs"></block>
            <block type="app"></block>
            <block type="var"></block>
            <block type="math_number"></block>
            <block type="math_arithmetic"></block>
        </xml>
    </div>
=======
    <div class="row">
        <div id="blocklyDiv" class="full-width"></div>
        <xml id="toolbox" style="display: none">
            <block type="functionDef"></block>
            <block type="functionCall"></block>
            <block type="abs"></block>
            <block type="app"></block>
            <block type="var"></block>
            <block type="math_number"></block>
            <block type="math_arithmetic"></block>
        </xml>
    </div>
    <div class="row">
        <a href="https://github.com/your_username/your_repository" target="_blank">View Source on GitHub</a>
    </div>
>>>>>>> updated
```

Please replace `your_username` and `your_repository` with your actual GitHub username and repository name. The `target="_blank"` attribute will open the link in a new tab.