aider: Added GitHub logo next to the link for viewing source on GitHub.
# Aider chat conversation:

USER: There is a nice github logo for this, the github cat, can you add it in?

ASSISTANT: Sure, we can add the GitHub logo next to the link. We will use an `img` tag for this. The `img` tag has a `src` attribute where we will put the URL of the GitHub logo.

Here is the change that needs to be made:

```html
Definitions/index.html
<<<<<<< HEAD
    <div class="row">
        <a href="https://github.com/your_username/your_repository" target="_blank">View Source on GitHub</a>
    </div>
=======
    <div class="row">
        <a href="https://github.com/your_username/your_repository" target="_blank">
            <img src="https://github.githubassets.com/images/modules/logos_page/GitHub-Mark.png" alt="GitHub Logo" width="20" height="20">
            View Source on GitHub
        </a>
    </div>
>>>>>>> updated
```

Please replace `your_username` and `your_repository` with your actual GitHub username and repository name. The `src` attribute of the `img` tag contains the URL of the GitHub logo. The `alt` attribute provides alternative text for the image if it cannot be displayed. The `width` and `height` attributes specify the dimensions of the image. Adjust these values as needed to fit your design.