aider: Preloaded a `functionDef` block into the workspace when the page loads.
# Aider chat conversation:

USER: I want to preload a certain block into the workspace so that the user has an easier start, how can I do that?

ASSISTANT: To preload a block into the workspace, you can use the `Blockly.Xml.domToWorkspace()` function. This function takes an XML representation of a block or set of blocks and adds them to the workspace.

You will need to modify the `lc-with-definitions/index.html` file.

Here are the steps:

1. Define the XML for the block you want to preload. This should be a string that follows the Blockly XML format. For example, if you wanted to preload a `functionDef` block, the XML might look like this: `<xml xmlns="https://developers.google.com/blockly/xml"><block type="functionDef" x="10" y="10"></block></xml>`. The `x` and `y` attributes specify the position of the block in the workspace.

2. After the workspace is injected, use `Blockly.Xml.domToWorkspace()` to add the block to the workspace. You will need to convert the XML string to a DOM object first, which you can do with `Blockly.utils.xml.textToDom()`.

Here is the *edit block* for this change:

```html
lc-with-definitions/index.html
<<<<<<< HEAD
        var workspace = Blockly.inject('blocklyDiv', {
            toolbox: document.getElementById('toolbox'),
            scrollbars: true
        });
=======
        var workspace = Blockly.inject('blocklyDiv', {
            toolbox: document.getElementById('toolbox'),
            scrollbars: true
        });

        var defaultBlocks = '<xml xmlns="https://developers.google.com/blockly/xml"><block type="functionDef" x="10" y="10"></block></xml>';
        var defaultBlocksDom = Blockly.utils.xml.textToDom(defaultBlocks);
        Blockly.Xml.domToWorkspace(defaultBlocksDom, workspace);
>>>>>>> updated
```

This will preload a `functionDef` block into the workspace when the page loads. You can replace the `defaultBlocks` string with the XML for any block or set of blocks you want to preload.