aider: Call `loadBlocksFromURL()` function after the workspace is created and default blocks are loaded in `lc-with-definitions/index.html`.
# Aider chat conversation:

USER: can we make it in such a way that the loadBlocksFromURL() is called automatically when navigating to a url that has blocks in the url?

ASSISTANT: Yes, we can achieve this by calling the `loadBlocksFromURL()` function immediately after the workspace is injected. This will ensure that if there are blocks in the URL, they will be loaded into the workspace automatically.

Here are the changes needed in the `lc-with-definitions/index.html` file:

```html
lc-with-definitions/index.html
<<<<<<< HEAD
        var workspace = Blockly.inject('blocklyDiv', {
            toolbox: document.getElementById('toolbox'),
            scrollbars: true
        });

        // No changes here, use defaultBlocks directly
        var defaultBlocksDom = Blockly.utils.xml.textToDom(defaultBlocks);
        Blockly.Xml.domToWorkspace(defaultBlocksDom, workspace);
=======
        var workspace = Blockly.inject('blocklyDiv', {
            toolbox: document.getElementById('toolbox'),
            scrollbars: true
        });

        // No changes here, use defaultBlocks directly
        var defaultBlocksDom = Blockly.utils.xml.textToDom(defaultBlocks);
        Blockly.Xml.domToWorkspace(defaultBlocksDom, workspace);

        // Load blocks from URL if present
        loadBlocksFromURL();
>>>>>>> updated
```

This change will call the `loadBlocksFromURL()` function immediately after the workspace is created and the default blocks are loaded. If there are blocks in the URL, they will replace the default blocks. If there are no blocks in the URL, the function will do nothing and the default blocks will remain.