aider: Refactored code to load default blocks only if there are no blocks in the URL.
# Aider chat conversation:

USER: can you modify the code so that in the case where we load a block from the URL we do not also load the default.

ASSISTANT: Yes, we can modify the code to load the default blocks only if there are no blocks in the URL. This can be achieved by moving the code that loads the default blocks into the `loadBlocksFromURL()` function, and only executing it if there are no blocks in the URL.

Here are the changes needed in the `lc-with-definitions/index.html` file:

```html
lc-with-definitions/index.html
<<<<<<< HEAD
        var workspace = Blockly.inject('blocklyDiv', {
            toolbox: document.getElementById('toolbox'),
            scrollbars: true
        });

        // No changes here, use defaultBlocks directly
        var defaultBlocksDom = Blockly.utils.xml.textToDom(defaultBlocks);
        Blockly.Xml.domToWorkspace(defaultBlocksDom, workspace);

        // Load blocks from URL if present
        loadBlocksFromURL();
=======
        var workspace = Blockly.inject('blocklyDiv', {
            toolbox: document.getElementById('toolbox'),
            scrollbars: true
        });

        // Load blocks from URL if present, else load default blocks
        loadBlocksFromURL();
>>>>>>> updated
```

And here are the changes needed in the `loadBlocksFromURL()` function:

```javascript
lc-with-definitions/index.html
<<<<<<< HEAD
        function loadBlocksFromURL() {
            var urlParams = new URLSearchParams(window.location.search);
            var encodedXml = urlParams.get('blocks');
            if (encodedXml) {
                var xmlText = atob(encodedXml);
                var xml = Blockly.utils.xml.textToDom(xmlText);
                Blockly.Xml.appendDomToWorkspace(xml, workspace);
            }
        }
=======
        function loadBlocksFromURL() {
            var urlParams = new URLSearchParams(window.location.search);
            var encodedXml = urlParams.get('blocks');
            if (encodedXml) {
                var xmlText = atob(encodedXml);
                var xml = Blockly.utils.xml.textToDom(xmlText);
                Blockly.Xml.appendDomToWorkspace(xml, workspace);
            } else {
                // No blocks in URL, use defaultBlocks
                var defaultBlocksDom = Blockly.utils.xml.textToDom(defaultBlocks);
                Blockly.Xml.domToWorkspace(defaultBlocksDom, workspace);
            }
        }
>>>>>>> updated
```

This change will load the default blocks only if there are no blocks in the URL. If there are blocks in the URL, they will be loaded into the workspace and the default blocks will not be loaded.