aider: Added collapsible section for instructions in index.html.
# Aider chat conversation:

USER: the instructions which are in the second div of index.html, can we make them fold in and out upon click? it could say sth like "Here for the first time?" and fold out upon click. what do you think? is there another nice way of doint sth like this?

ASSISTANT: Sure, we can use a simple HTML details and summary tag to create a collapsible section. The details tag creates a disclosure widget in which information is visible only when the widget is toggled into an "open" state. A summary tag can be used as the first child of the details element to provide a label for the disclosure widget.

Here are the changes we need to make:

1. Modify the `lc-with-definitions/index.html` file.

```html
lc-with-definitions/index.html
<<<<<<< HEAD
    <div>
        <p> The default BlocklyLambda program you see below defines the (anonymous) function "plus one" ($\lambda x.x+1$) and applies it to the number 2. </p>
        <p> "Generate JavaScript" compiles this to JavaScript (which is a programming language that runs in your browser) and "Run JavaScript" executes it.</p>
        <p> You can remove the default program by dragging it out of the workspace. You can write your own program by dragging blocks from the toolbox to the workspace. You can save your program to a file and load it from a file. </p>
        <p> You can also save your program to a URL and load it from a URL. This is useful if you want to share your program with others. </p>
    </div>
=======
    <details>
        <summary>Here for the first time?</summary>
        <p> The default BlocklyLambda program you see below defines the (anonymous) function "plus one" ($\lambda x.x+1$) and applies it to the number 2. </p>
        <p> "Generate JavaScript" compiles this to JavaScript (which is a programming language that runs in your browser) and "Run JavaScript" executes it.</p>
        <p> You can remove the default program by dragging it out of the workspace. You can write your own program by dragging blocks from the toolbox to the workspace. You can save your program to a file and load it from a file. </p>
        <p> You can also save your program to a URL and load it from a URL. This is useful if you want to share your program with others. </p>
    </details>
>>>>>>> updated
```

This will create a collapsible section with the title "Here for the first time?". When the user clicks on the title, the section will expand to show the instructions. When the title is clicked again, the section will collapse.