aider: Added change listener to automatically update the URL when blocks are changed.
# Aider chat conversation:

USER: can you help now?

ASSISTANT: Of course! Please provide the details of the changes you want to make to the code.

USER: there may be problems with `saveBlocksToURL` when I change the blocks the URL does not change

ASSISTANT: The `saveBlocksToURL` function in your `lc-with-definitions/index.html` file is designed to save the current state of the blocks to the URL. However, it does not automatically update the URL when you change the blocks. It only updates the URL when you manually call the `saveBlocksToURL` function, for example by clicking a button.

If you want the URL to automatically update whenever you change the blocks, you could add an event listener to the Blockly workspace that calls `saveBlocksToURL` whenever a change event occurs.

Here is how you can modify the code:

```html
lc-with-definitions/index.html
<<<<<<< HEAD
        var workspace = Blockly.inject('blocklyDiv', {
            toolbox: document.getElementById('toolbox'),
            scrollbars: true
        });

        // Load blocks from URL if present, else load default blocks
        loadBlocksFromURL();
=======
        var workspace = Blockly.inject('blocklyDiv', {
            toolbox: document.getElementById('toolbox'),
            scrollbars: true
        });

        // Add event listener to workspace
        workspace.addChangeListener(saveBlocksToURL);

        // Load blocks from URL if present, else load default blocks
        loadBlocksFromURL();
>>>>>>> updated
```

This change adds a change listener to the Blockly workspace. The `saveBlocksToURL` function is called whenever a change event occurs in the workspace, which means the URL will be updated whenever you change the blocks.