aider: Refactored `loadBlocksFromURL` function to add error handling and separate loading default blocks into a separate function.
# Aider chat conversation:

USER: there are problems with load Blocks from URL ... sometimes it loads sth and sts not

ASSISTANT: The `loadBlocksFromURL` function in your `lc-with-definitions/index.html` file is designed to load blocks from the URL. It does this by looking for a URL parameter named `blocks`, which should contain the blocks encoded as a base64 string.

If the `blocks` parameter is not present in the URL, or if its value is not a valid base64 string, or if the decoded string is not valid XML, then the function will not be able to load the blocks. In these cases, it falls back to loading the default blocks.

If you're having trouble with this function, it could be due to one of these issues. To help diagnose the problem, you could add some error handling code to the function. This would allow you to see more detailed error messages when something goes wrong.

Here is how you can modify the code:

```html
lc-with-definitions/index.html
<<<<<<< HEAD
        function loadBlocksFromURL() {
            var urlParams = new URLSearchParams(window.location.search);
            var encodedXml = urlParams.get('blocks');
            if (encodedXml) {
                var xmlText = atob(encodedXml);
                var xml = Blockly.utils.xml.textToDom(xmlText);
                Blockly.Xml.appendDomToWorkspace(xml, workspace);
            } else {
                // No blocks in URL, use defaultBlocks
                var defaultBlocksDom = Blockly.utils.xml.textToDom(defaultBlocks);
                Blockly.Xml.domToWorkspace(defaultBlocksDom, workspace);
            }
        }
=======
        function loadBlocksFromURL() {
            var urlParams = new URLSearchParams(window.location.search);
            var encodedXml = urlParams.get('blocks');
            if (encodedXml) {
                try {
                    var xmlText = atob(encodedXml);
                    var xml = Blockly.utils.xml.textToDom(xmlText);
                    Blockly.Xml.appendDomToWorkspace(xml, workspace);
                } catch (e) {
                    console.error('Error loading blocks from URL:', e);
                    loadDefaultBlocks();
                }
            } else {
                loadDefaultBlocks();
            }
        }

        function loadDefaultBlocks() {
            var defaultBlocksDom = Blockly.utils.xml.textToDom(defaultBlocks);
            Blockly.Xml.domToWorkspace(defaultBlocksDom, workspace);
        }
>>>>>>> updated
```

This change adds a try-catch block around the code that decodes the base64 string and loads the blocks. If an error occurs, it logs the error to the console and then loads the default blocks. It also moves the code for loading the default blocks into a separate function for clarity.